separate different components into different packages
