+++ helpers/helpers.go
@@ -1,0 +1,17 @@
+package helpers
+// https://www.dotnetperls.com/duplicates-go
+func RemoveDuplicatesUnordered(elements []string) []string {
+    encountered := map[string]bool{}
+
+    // Create a map of all unique elements.
+    for v:= range elements {
+        encountered[elements[v]] = true
+    }
+
+    // Place all keys from the map into a slice.
+    result := []string{}
+    for key, _ := range encountered {
+        result = append(result, key)
+    }
+    return result
+}
