+++ structs/structs.go
@@ -1,0 +1,7 @@
+package structs
+
+type CommitInfo struct {
+	IsMergeCommit bool
+	Author     string
+	Date     string
+}
