+++ tests/environments/chef-uploader-test-env.rb
@@ -1,0 +1,8 @@
+name 'chef-uploader-test-env'
+description 'This is a test environment created by Chef Uploader.'
+
+cookbook_versions({
+})
+
+default_attributes()
+
