+++ tests/roles/chef-uploader-test-role.rb
@@ -1,0 +1,3 @@
+name 'chef-uploader-test-role'
+run_list()
+default_attributes()
